pyproject.toml
# Sumii Mobile API - Python Project Configuration (uv)

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sumii-mobile-api"
version = "0.1.0"
description = "Sumii Mobile API - Backend for Mobile App (User-Facing)"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.17.0",
    "psycopg2-binary>=2.9.0",
    "mistralai>=1.0.0",
    "mcp>=1.0.0",  # Required for Mistral RunContext
    "griffe>=1.0.0",  # Required for Mistral RunContext function tools
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.0.0",
    "weasyprint>=61.0",
    "markdown>=3.5.0",  # For markdown to HTML conversion
    "boto3>=1.34.0",
    "email-validator>=2.0.0",
    "greenlet>=3.0.0",  # Required for SQLAlchemy async operations
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
    "httpx>=0.27.0",  # For async test client
    "requests>=2.32.0",  # For integration tests
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["B008"]  # Allow FastAPI Depends() in function defaults

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "^alembic/",
    "^build/",
    "^dist/"
]

[[tool.mypy.overrides]]
module = ["alembic.*", "requests.*", "markdown.*"]
ignore_missing_imports = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["infrastructure*"]
